requirements.txt
pyserial>=3.5
PyVISA>=1.12.0
visa>=1.0.0
